backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyjwt==2.8.0
chardet==5.2.0
